[tests] Update node tests to run concurrently
