+++ packages/node/package.json
@@ -11,9 +11,9 @@
   },
   "scripts": {
     "build": "node build",
-    "test-integration-once": "yarn test test/integration.test.js",
     "test": "jest --env node --verbose --bail --runInBand",
     "test-unit": "yarn test test/prepare-cache.test.ts test/utils.test.ts",
+    "test-integration-once": "yarn test test/integration-*.test.js",
     "prepublishOnly": "node build"
   },
   "files": [
