+++ packages/node/test/integration-3.test.js
@@ -1,0 +1,1 @@
+require('./integration-setup')(3);
