+++ packages/node/test/integration-4.test.js
@@ -1,0 +1,1 @@
+require('./integration-setup')(4);
